moved some stuff around a bit
I think this will make it easier for navigation! Maybe.
