+++ services/dataHandler.js
@@ -1,5 +1,7 @@
 export async function getAvailablePlayers() {
     //in the future we will use fetch() to grab this from the server, or use websockets or something!
-    let response = Math.floor(Math.random() * 6)
-    return response
+    const names = ['Callisto', 'Lotus', 'inS', 'sommie', 'dmlr_']
+    let howMany = Math.floor(Math.random() * 6)
+    let randomList = names.sort(() => Math.random() > 0.5)
+    return randomList.slice(0,howMany)
 }
